Items entity relocated to /entities folder
