+++ src/items/entities/items.entity.ts
