+++ src/items/entities/items.interface.ts
